my first commit

+++ g23-boutique-master/src/main/java/za/ac/cput/domain/Payment.java
@@ -69,7 +69,7 @@ public class Payment {
         }
 
         public Builder setAmount(long amount) {
-            this.amount = amount;
+            this.amount = (double) amount;
             return this;
         }
